Refactor maze generation: move MazeGenerator to the correct package and remove unused import in MazeModel
